+++ blueprints/flexberry-enum/files/__root__/transforms/__name__.js
@@ -2,5 +2,6 @@ import FlexberryEnum from 'ember-flexberry-data/transforms/flexberry-enum';
 import <%= className %>Enum from '../enums/<%= name %>';
 
 export default FlexberryEnum.extend({
-  enum: <%= className %>Enum
+  enum: <%= className %>Enum,
+  sourceType: '<%= sourceType %>'
 });